% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch_UMAP_Plot.R
\name{Batch_UMAP_Plot}
\alias{Batch_UMAP_Plot}
\title{Title}
\usage{
Batch_UMAP_Plot(
  combined_dataset,
  point_size = 0.001,
  cell_type = "seurat_pbmc_type"
)
}
\arguments{
\item{cell_type}{}
}
\value{

}
\description{
Title
}
