% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchDonors_DataAvailablity_Dotsplots.R
\name{BatchDonors_DataAvailablity_Dotsplots}
\alias{BatchDonors_DataAvailablity_Dotsplots}
\title{Title}
\usage{
BatchDonors_DataAvailablity_Dotsplots(
  rna_dataframe,
  meta_data = "sample.visitName"
)
}
\arguments{
\item{meta_data}{}
}
\value{

}
\description{
Title
}
