% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_Density.R
\name{Calculate_Density}
\alias{Calculate_Density}
\title{Title}
\usage{
Calculate_Density(
  Combined_dataset,
  group = c("sample.visitName"),
  n_cells_per_group = 1e+05
)
}
\arguments{
\item{n_cells_per_group}{}
}
\value{

}
\description{
Title
}
