% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellNumber_Batch_Boxplot.R
\name{CellNumber_Batch_Boxplot}
\alias{CellNumber_Batch_Boxplot}
\title{Title}
\usage{
CellNumber_Batch_Boxplot(
  combined_dataset,
  group = "batch_id",
  sample_id = "pbmc_sample_id",
  plotly = "False"
)
}
\arguments{
\item{plotly}{}
}
\value{

}
\description{
Title
}
