% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellType_Composition_Heatmap.R
\name{CellType_Composition_Heatmap}
\alias{CellType_Composition_Heatmap}
\title{Title}
\usage{
CellType_Composition_Heatmap(
  combined_dataset,
  cell_type = "seurat_pbmc_type",
  group = "pbmc_sample_id",
  Sort = "False"
)
}
\arguments{
\item{group}{}
}
\value{

}
\description{
Title
}
