% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileDescToDataframe_beta.R
\name{fileDescToDataframe_beta}
\alias{fileDescToDataframe_beta}
\title{Title}
\usage{
fileDescToDataframe_beta(descriptors, keep_labs = FALSE)
}
\arguments{
\item{keep_labs}{}
}
\value{

}
\description{
Title
}
