% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge_Dataset.R
\name{Merge_Dataset}
\alias{Merge_Dataset}
\title{Title}
\usage{
Merge_Dataset(rna_dataframe, filePath = "filePath", Bridging_Control = "FALSE")
}
\arguments{
\item{Bridging_Control}{}
}
\value{

}
\description{
Title
}
