% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMAP_Scatter_Plot.R
\name{UMAP_Scatter_Plot}
\alias{UMAP_Scatter_Plot}
\title{Title}
\usage{
UMAP_Scatter_Plot(
  combined_dataset,
  group = "seurat_pbmc_type",
  dot_alpha = 0.3,
  dot_size = 0.001,
  plot_width = 1000,
  plot_height = 800
)
}
\arguments{
\item{plot_height}{}
}
\value{

}
\description{
Title
}
