% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneDistribution_Violin_Plot.R
\name{GeneDistribution_Violin_Plot}
\alias{GeneDistribution_Violin_Plot}
\title{Title}
\usage{
GeneDistribution_Violin_Plot(combined_dataset, plotly = "False")
}
\arguments{
\item{plotly}{}
}
\value{

}
\description{
Title
}
